+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3
+
+* Added hf_transfer support
+* Fix bug in getting file length from HTTP response
+* Added way to throw on error for only specific initialization errors
+
 ## 0.0.2
 
 * Added Xet storage support